revamps the visual style of this config
removes rounded corners and gaps between windows,
and the top bar is no longer translucent. Instead,
it now checks a pixel right below the middle of
the bar (that pixel should be part of the
background or the window titlebar) using my
`pixgrab` custom package (couldn't find another
on `nixpkgs` or even a script which executes
faster than my solution. This needs to run at
tiny intervals for seamless color transitions).

TODO: maybe create a new variant of `pixgrab`
which continously outputs the pixel color without
terminating, and use `deflisten` instead of
`defpoll` in the `eww` bar, so the program
is no longer opened and closed so often, which
is quite inefficient. (didn't measure it though,
maybe it's not such a big deal)

Signed-off-by: RCX777 <robertcnst02@gmail.com>